requirements.txt
PyMuPDF>=1.22.3
fitz>=0.0.1
Pillow>=2.1.0